pyproject.toml
[project]
name = "obsidian-mcp"
version = "0.1.0"
description = "MCP server for read-only Obsidian vault access"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "python-dotenv",
    "python-frontmatter",
]

[project.scripts]
obsidian-mcp = "obsidian_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_mcp"]
